requirements.txt
python_version >= "3.8"
ultralytics==8.2.77
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
streamlit==1.37.1
streamlit-webrtc==0.47.7
pandas==2.2.2
openpyxl==3.1.5
pyyaml==6.0.2
joblib==1.4.2
numpy==1.26.4
scipy==1.11.4
pillow==10.3.0
screeninfo==0.8.1
albumentations==1.4.8
roboflow==1.1.25
tqdm==4.66.4
